+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ## [Unreleased](https://gitlab.heigit.org/climate-action/climatoology/-/compare/7.0.0rc3...main)
 
+### Fix
+
+- a problem where computations with artifacts where some had a description and others didn't
+  failed ([#260](https://gitlab.heigit.org/climate-action/climatoology/-/issues/260)).
+
 ## [7.0.0rc3](https://gitlab.heigit.org/climate-action/climatoology/-/compare/7.0.0rc3...main)
 
 ### Fix